Add files via upload
This code creates a variety plots and graphs of the "titanic" data set using the "ggplot2" package. Multiple variables and styles of graphs are looked at. Not all the graphs show/represent important findings. The main purpose of this assignment was to get familiar with data visualization options with "ggplot2".