Add files via upload
This assignment again goes through the model building process for both the "titanic" and "adult" data sets. Instead of splitting the data into training and testing data to measure the accuracy though, cross validation is used. Multiple combinations of different hyper-parameters are used when training the model as well which makes this code take more time to run. A random forest is also trained in this code.