Assignment 17- Case Study
In this assignment, I go through the full data mining process of imputing missing values, manipulating variables, visualization data, balancing data, and finally building multiple models. The data set used in this study is a car accident data set where the goal is to predict the binary target variable "fatal" or "not fatal".